requirements.txt
fastapi
uvicorn
gunicorn
guardrails-ai